# docs/openapi-rest.yaml
openapi: 3.0.3
info:
  title: LCP Platform API
  version: 1.1.0
  description: >
    REST and GraphQL surface for the self-hosted LCP platform. This document
    covers the admin dashboard login flow, publication processing, license
    lookup, status APIs, content downloads, publication catalog operations,
    runtime endpoints, and the GraphQL operations used by integration clients
    such as WordPress.
servers:
  - url: https://yourdomain.com
  - url: https://staging.yourdomain.com
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: Sign in and mint a JWT for the admin dashboard or publisher flow
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/publications:
    get:
      summary: List publication catalog entries
      operationId: listPublications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Publication catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a publication record and optionally process a file
      operationId: createPublication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationRequest"
      responses:
        "201":
          description: Publication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/lcp/process:
    post:
      summary: Process publication input data or an uploaded publication
      operationId: processPublication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
      responses:
        "200":
          description: Processing completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/lcp/status:
    get:
      summary: Get service or process status
      operationId: getProcessStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Status payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StatusEnvelope"
                  - $ref: "#/components/schemas/ProcessStatus"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/users:
    get:
      summary: List admin-visible users
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - name: X-2FA-Code
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/licenses:
    get:
      summary: List licenses for the admin dashboard
      operationId: listAdminLicenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLicensesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/licenses/{licenseId}:
    patch:
      summary: Update a license loan expiry date
      operationId: patchAdminLicense
      security:
        - bearerAuth: []
      parameters:
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLicensePatchRequest"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLicense"
...
  /licenses/{licenseId}/status:
    get:
      summary: Readium license status document
      operationId: getLicenseStatusDocument
      security: []
      parameters:
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Readium license status document
          content:
            application/vnd.readium.license.status.v1.0+json:
              schema:
                $ref: "#/components/schemas/LicenseStatusDocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /graphql:
    post:
      summary: GraphQL endpoint for publication and license workflows
      operationId: graphQLOperation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GraphQLRequest"
      responses:
        "200":
          description: GraphQL response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphQLResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Role or 2FA validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      additionalProperties: false
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: toghyani
        password:
          type: string
          example: e40949a95921fb6e
        twoFactor:
          type: string
          example: 65b5ec
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          example: admin
        subject:
          type: string
          example: toghyani
        expiresAt:
          type: string
          format: date-time
    Publication:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        language:
          type: string
        subjects:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        right_print:
          type: integer
          description: Allowed print count for the publication; 0 disables print rights.
        right_copy:
          type: integer
          description: Allowed copy count for the publication; 0 disables copy rights.
        status:
          type: string
          example: active
        file_path:
          type: string
        encrypted_path:
          type: string
        encrypted_uri:
          type: string
        checksum:
          type: string
        licenseDurationDays:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicationListResponse:
      type: object
      properties:
        publications:
          type: array
          items:
            $ref: "#/components/schemas/Publication"
    PublicationRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        language:
          type: string
        subjects:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        right_print:
          type: integer
          description: Allowed print count for the publication; 0 disables print rights.
        right_copy:
          type: integer
          description: Allowed copy count for the publication; 0 disables copy rights.
        encrypted_uri:
          type: string
        checksum:
          type: string
        license_duration_days:
          type: integer
        file:
          type: string
          description: Base64 encoded content of the uploaded publication
        status:
          type: string
          description: Publication status for catalog management
...
    ProcessRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Example Publication
        file:
          type: string
          description: Base64 encoded publication content selected in the browser
        data:
          type: object
          description: Optional opaque metadata payload for future extensions
    ProcessStatus:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          example: completed
        publicationId:
          type: string
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StatusEnvelope:
      type: object
      properties:
        status:
          type: string
          example: ok
        uptimeSec:
          type: integer
          format: int64
        processes:
          type: array
          items:
            $ref: "#/components/schemas/ProcessStatus"
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
    AdminLicense:
      type: object
      properties:
        id:
          type: string
        publicationID:
          type: string
        userID:
          type: string
        passphrase:
          type: string
        hint:
          type: string
        publicationURL:
          type: string
        rightPrint:
          type: integer
        rightCopy:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminLicensesResponse:
      type: object
      properties:
        licenses:
          type: array
          items:
            $ref: "#/components/schemas/AdminLicense"
    AdminLicensePatchRequest:
      type: object
      properties:
        endDate:
          type: string
          format: date-time
...
    GraphQLRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: >
            Supported operations include uploadPublication, createLicense,
            revokeLicense, licenses, and publications.
        variables:
          type: object
          additionalProperties: true
    GraphQLResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string